update Modal metadata styles and remove formik component
