+++ src/components/AstroUpload.tsx
@@ -1,4 +1,3 @@
-import AstroUploadF from "./AstroUploadF";
 import AstroUploadForm from "./AstroUploadForm";
 
 const AstroUpload = () => {
@@ -7,12 +6,11 @@ const AstroUpload = () => {
       <div>
         <h1 className="font-bold text-5xl">Create your Astro NFT </h1>
         <p className="text-neutral-300 my-2">
-          Once your NFT is created you won’t be able to update or change any of
-          this <br />
+          Once your NFT is created you will not be able to update or change any of this <br />
           information.
         </p>
       </div>
-      <AstroUploadF />
+      <AstroUploadForm />
     </div>
   );
 };
