+++ src/components/AstroUploadForm.tsx
@@ -1,45 +1,82 @@
 "use client";
 
-import { useRef, useState } from "react";
-import Image from "next/image";
-import { ArrayHelpers, FieldArray, Form, Formik, FormikHelpers } from "formik";
-
-import { MetadataNFT, NFT, TraitNFT } from "../types/types";
-import { MdOutlineFileUpload } from "react-icons/md";
-import { twMerge } from "tailwind-merge";
+import React, { memo, useEffect, useMemo, useRef } from "react";
+import { Address, ProviderRpcError, TransactionExecutionError } from "viem";
+import { useFieldArray, useForm } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod";
 import toast from "react-hot-toast";
-import { Address, TransactionExecutionError } from "viem";
-import Input from "./Form/Input";
-import Textarea from "./Form/Textarea";
-import Trait from "./Trait";
+import Image from "next/image";
+import { z } from "zod";
+
+import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from "./ui/form";
+import { Input } from "./ui/input";
+import { Textarea } from "./ui/textarea";
+import { Button } from "./ui/button";
+
+import Modal from "./Modal";
+import { Trait } from "./Trait2";
+import { newTrait } from "@/constants/constants";
+import { NFTForm as NFTFormSchema } from "@/schema/nft";
+import { cn } from "@/lib/utils";
+import { MdOutlineFileUpload } from "react-icons/md";
 import { uploadToPinata } from "@/actions/pinataUpload";
 import { useWriteAstroNft } from "@/nftMarketHooks";
-import { newTrait } from "@/constants/constants";
-import Clipboard from "./Clipboard";
-import Modal from "./Modal";
-import AstroCard from "./AstroCard";
-import Button from "./Button";
-import { NFTSchema } from "@/schema/metadataNFTSchema";
 import { config } from "../../config";
-import AstroUploadF from "./AstroUploadF.1";
+import AstroCard from "./AstroCard";
+import Clipboard from "./Clipboard";
+import SubmitButton from "@/app/nft/mint/_components/SubmitButton";
 
-const initialMetadata: NFT = {
+type NFTForm = z.infer<typeof NFTFormSchema>;
+const initialMetadata: NFTForm = {
   description: "",
   name: "",
   traits: [],
   file: null,
 };
 
-const AstroUploadForm = () => {
+const AstroUploadF = () => {
   const nftFileRef = useRef<HTMLInputElement>(null);
+
+  const form = useForm<NFTForm>({
+    defaultValues: initialMetadata,
+    resolver: zodResolver(NFTFormSchema),
+    mode: "all",
+  });
+
+  const {
+    append,
+    fields: traits,
+    remove,
+  } = useFieldArray({
+    control: form.control,
+    name: "traits",
+    shouldUnregister: true,
+  });
+
   const {
     writeContractAsync: writeAstro,
     data: txHash,
     isSuccess,
-    reset,
+    reset: astroReset,
+    isPending,
   } = useWriteAstroNft();
-  const [nftFile, setNftFile] = useState<File | null>(null);
-  const [metadata, setMetadata] = useState<MetadataNFT | null>(null);
+
+  // const [formActionState, onFormAction] = useFormState(uploadToPinata, {
+  //   status: "default",
+  // });
+  //derive
+
+  const { isSubmitting, isDirty, isValid } = form.formState;
+  const unableToSubmit = isPending || isSubmitting || !isValid || !isDirty;
+
+  const { file, description, name } = form.getValues();
+
+  const imageURI = useMemo(() => {
+    return file ? URL.createObjectURL(file) : null;
+  }, [file]);
+
+  // console.log({ imageURI });
+  //event handlers
 
   const handleBrowseFile = () => {
     if (!nftFileRef.current) return;
@@ -48,212 +85,199 @@ const AstroUploadForm = () => {
   };
 
   const handleContractReset = () => {
-    reset();
-    setMetadata(null);
-    setNftFile(null);
+    astroReset();
+    form.reset();
   };
 
-  const handleSubmit = async (
-    values: NFT,
-    { setSubmitting, resetForm }: FormikHelpers<NFT>
-  ) => {
+  const handleAction = async () => {
     try {
-      setSubmitting(true);
-      const data = new FormData();
-      if (values.file) {
-        data.append("file", values.file, values.file.name);
+      const formData = new FormData();
+      formData.append("name", name);
+      formData.append("description", description);
+      formData.append("file", file!);
+      const traits = form.getValues("traits");
+      if (traits && traits.length > 0) {
+        formData.append("traits", JSON.stringify(traits));
       }
 
-      const metadata: MetadataNFT = { ...values };
-
-      data.append("metadata", JSON.stringify(metadata));
-      const response = await uploadToPinata(data);
-      setMetadata(response.metadata);
-
-      toast
-        .promise(
-          writeAstro({
-            address: config.pubNftMarketAddress as Address,
-            functionName: "mint",
-            args: [response.cid],
-          }),
-          {
-            loading: `Minting ${response.metadata.name}`,
-            success: "Your NFT has been mintent Successfully",
-            error: "Error Minting your NFT",
-          }
-        )
-        .then(() => {
-          resetForm();
-        });
-    } catch (error) {
-      console.log({ error });
-      if (error instanceof TransactionExecutionError) {
-        toast.error(error.shortMessage, { style: { color: "#fa594d" } });
+      const { status, tokenURI, metadata, formError, errors } = await uploadToPinata(formData);
+      if (status === "error") {
+        if (errors) {
+          toast.error(errors, { style: { color: "#fa594d" } });
+          return;
+        }
+        if (formError) {
+          Object.keys(formError).forEach((key) => {
+            if (!formError) return;
+            const message = formError[key as keyof NFTForm];
+            form.setError(key as keyof NFTForm, {
+              message,
+            });
+          });
+          return;
+        }
         return;
       }
 
-      if (error instanceof Error)
-        toast.error(error.message, { style: { color: "#fa594d" } });
-    } finally {
-      setSubmitting(false);
+      toast.promise(
+        writeAstro({
+          address: config.pubNftMarketAddress as Address,
+          functionName: "mint",
+          args: [tokenURI!],
+        }),
+        {
+          loading: `Minting ${metadata!.name}`,
+          success: "Your NFT has been mintent Successfully",
+          error: (err: ProviderRpcError) => err.shortMessage,
+        }
+      );
+    } catch (error) {
+      console.error({ error });
+      if (error instanceof Error) toast.error(error.message, { style: { color: "#fa594d" } });
     }
   };
 
   return (
-    <>
-      <section id="upload-nft" className="grid grid-cols-2 gap-x-12 pl-8">
-        <div className="min-h-fit pt-12">
-          <Formik
-            initialValues={initialMetadata}
-            onSubmit={handleSubmit}
-            validationSchema={NFTSchema}
-          >
-            {({ values, setFieldValue, isSubmitting, isValid, dirty }) => {
-              return (
-                <Form className="space-y-5">
-                  <Input
-                    ref={nftFileRef}
-                    type="file"
-                    name="file"
-                    className="hidden"
-                    accept="image/*"
-                    value={""}
-                    onChange={(e) => {
-                      if (!e.currentTarget.files) return;
-
-                      setFieldValue("file", e.currentTarget.files[0], true);
-                      setNftFile(e.currentTarget.files[0]);
-                    }}
-                    withMessage={false}
-                    disabled={isSubmitting}
-                  />
-                  <Input
-                    type="text"
-                    name="name"
-                    placeholder="Name your NFT"
-                    label="Name: *"
-                    disabled={isSubmitting}
-                  />
-                  <Textarea
-                    name="description"
-                    placeholder="Describe your NFT"
-                    label="Description: *"
-                    className="resize-none scroll-w-0"
-                    rows={4}
-                    disabled={isSubmitting}
-                  />
-                  <div>
-                    <p className="font-semibold">Traits</p>
-                    <span className="text-neutral-300">
-                      Traits describe attributes of your item. They appear as
-                      filters inside your collection page and are also listed
-                      out inside your item page.
-                    </span>
-                    <FieldArray
-                      name="traits"
-                      render={({ remove, push }: ArrayHelpers<TraitNFT[]>) => (
-                        <div className="my-2">
-                          <div className="grid grid-cols-1 gap-2">
-                            {values.traits?.map((trait, idx) => (
-                              <Trait
-                                key={idx}
-                                idx={idx}
-                                onDelete={remove}
-                                disabled={isSubmitting}
-                              />
-                            ))}
-                          </div>
-                          <button
-                            type="button"
-                            className="font-semibold hover:text-neutral-400"
-                            onClick={() => {
-                              push(newTrait);
-                            }}
-                          >
-                            + Add trait
-                          </button>
-                        </div>
-                      )}
+    <section id="upload-nft" className="grid grid-cols-2 gap-x-12 pl-8">
+      <div className="min-h-fit pt-12">
+        <Form {...form}>
+          <form className="flex flex-col gap-y-4" action={handleAction}>
+            <input
+              type="file"
+              {...form.register("file")}
+              ref={nftFileRef}
+              className="hidden"
+              accept="image/*"
+              onChange={(e) => {
+                if (!e.currentTarget.files) return;
+                form.setValue("file", e.currentTarget.files[0], {
+                  shouldValidate: true,
+                });
+              }}
+              disabled={isSubmitting}
+            />
+            <FormField
+              control={form.control}
+              name="name"
+              render={({ field, fieldState: { error } }) => (
+                <FormItem>
+                  <FormLabel className="font-semibold">Name: *</FormLabel>
+                  <FormControl>
+                    <Input
+                      {...field}
+                      placeholder="Astro NFT name"
+                      className={error && "border-red-500 focus-visible:border-red-500"}
+                      maxLength={30}
+                      disabled={isSubmitting}
                     />
-                  </div>
-                  <button
-                    disabled={isSubmitting || !isValid || !dirty}
-                    type="submit"
-                    className="bg-orange-500 py-2 px-8 rounded-md text-xl font-semibold disabled:opacity-50 disabled:cursor-not-allowed"
-                  >
-                    Create
-                  </button>
-                </Form>
-              );
-            }}
-          </Formik>
-        </div>
-        <div className="p-2 ">
-          <div
-            className={twMerge(
-              `relative overflow-hidden aspect-square flex flex-col items-center justify-center rounded-lg hover:bg-neutral-500/10 hover:cursor-pointer transition`,
-              !nftFile && " border border-neutral-500 border-dashed"
-            )}
-            onClick={handleBrowseFile}
-          >
-            {nftFile ? (
-              <Image
-                src={URL.createObjectURL(nftFile)}
-                fill
-                className="object-cover hover:opacity-80 transition"
-                alt={nftFile.name}
-              />
-            ) : (
-              <>
-                <MdOutlineFileUpload size={40} />
-                <p className="text-xs  text-center">
-                  <span className="font-semibold text-sky-500 text-sm ">
-                    Browse files
-                  </span>
-                  <br />
-                  Max size: 50MB
-                  <br />
-                  JPG, JPEG, PNG, SVG or GIF
-                </p>
-              </>
-            )}
-          </div>
-        </div>
-        <Modal
-          isOpen={isSuccess && Boolean(metadata)}
-          onClose={handleContractReset}
-        >
-          <AstroCard
-            image={nftFile ? URL.createObjectURL(nftFile) : ""}
-            imageDescripion={"image"}
-          >
-            <div className="space-y-[4px]">
-              <h2 className="font-semibold first-letter:uppercase">
-                {metadata?.name}
-              </h2>
-              <p className="font-thin">{metadata?.description}</p>
-              <div>
-                <p>Transaction at:</p>
-                <Clipboard copy={txHash || ""}>
-                  <p className="break-all font-thin">
-                    <b>{txHash} </b>
-                  </p>
-                </Clipboard>
-              </div>
-              <div className="flex gap-x-2 font-medium">
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+            <FormField
+              control={form.control}
+              name="description"
+              render={({ field, fieldState: { error } }) => (
+                <FormItem>
+                  <FormLabel className="font-semibold">Description: *</FormLabel>
+                  <FormControl>
+                    <Textarea
+                      {...field}
+                      placeholder="Describe your NFT"
+                      className={error && "border-red-500 focus-visible:border-red-500"}
+                      rows={4}
+                      maxLength={150}
+                      disabled={isSubmitting}
+                    />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+            <div>
+              <p className="font-semibold">Traits</p>
+              <span className="text-neutral-300">
+                Traits describe attributes of your item. They appear as filters inside your
+                collection page and are also listed out inside your item page.
+              </span>
+              <div className="my-2 space-y-2">
+                {traits.map((trait, idx) => (
+                  <Trait
+                    key={trait.id}
+                    control={form.control}
+                    idx={idx}
+                    onDelete={remove}
+                    disabled={isSubmitting}
+                  />
+                ))}
                 <Button
-                  text="List on Market"
-                  className="bg-orange-500 text-white "
-                />
-                <Button text="Gallery" />
+                  variant={"ghost"}
+                  size={"sm"}
+                  type="button"
+                  className="font-semibold "
+                  onClick={() => {
+                    append(newTrait);
+                  }}
+                >
+                  + Add trait
+                </Button>
               </div>
+              <SubmitButton disabled={unableToSubmit} />
             </div>
-          </AstroCard>
-        </Modal>
-      </section>
-    </>
+          </form>
+        </Form>
+      </div>
+
+      <div className="p-2">
+        <div
+          className={cn(
+            `relative overflow-hidden aspect-1 flex flex-col items-center justify-center rounded-lg hover:bg-neutral-500/10 hover:cursor-pointer transition`,
+            !file && " border border-neutral-500 border-dashed"
+          )}
+          onClick={handleBrowseFile}
+        >
+          {imageURI ? (
+            <Image
+              src={imageURI}
+              fill
+              className="object-cover hover:opacity-80 transition"
+              alt={"your nft image"}
+            />
+          ) : (
+            <>
+              <MdOutlineFileUpload size={40} />
+              <p className="text-xs  text-center">
+                <span className="font-semibold text-sky-500 text-sm ">Browse files</span>
+                <br />
+                Max size: 50MB
+                <br />
+                JPG, JPEG, PNG, SVG or GIF
+              </p>
+            </>
+          )}
+        </div>
+      </div>
+      <Modal isOpen={isSuccess} onClose={handleContractReset}>
+        <AstroCard image={imageURI} imageDescripion={"image"}>
+          <div className="space-y-[4px]">
+            <h2 className="text-xl font-semibold first-letter:uppercase">{name}</h2>
+            <p className="font-thin">{description}</p>
+            <div className="">
+              <h3>Transaction:</h3>
+              <Clipboard copy={txHash!}>
+                <p className="break-all text-sm font-semibold">{txHash}</p>
+              </Clipboard>
+            </div>
+            <div className="flex gap-x-2 font-medium">
+              <Button>List on Market</Button>
+              <Button variant={"secondary"}>Gallery</Button>
+            </div>
+          </div>
+        </AstroCard>
+      </Modal>
+    </section>
   );
 };
 
-export default AstroUploadForm;
+export default AstroUploadF;
