+++ src/components/Clipboard.tsx
@@ -1,3 +1,4 @@
+import { cn } from "@/lib/utils";
 import React, { ReactNode } from "react";
 import toast from "react-hot-toast";
 import { BiSolidCopyAlt } from "react-icons/bi";
@@ -10,12 +11,7 @@ type ClipboardProps = {
   btnClassName?: string;
 };
 
-const Clipboard = ({
-  className,
-  children,
-  copy,
-  btnClassName,
-}: ClipboardProps) => {
+const Clipboard = ({ className, children, copy, btnClassName }: ClipboardProps) => {
   const handleCopyToClipboard = () => {
     navigator.clipboard.writeText(copy);
     toast.success("Copied", {
@@ -23,21 +19,10 @@ const Clipboard = ({
     });
   };
   return (
-    <div
-      className={twMerge(
-        `flex justify-between gap-x-2 text-green-400 bg-green-400/10 rounded-lg p-1 break-all `,
-        className
-      )}
-    >
+    <div className={cn(`flex justify-between gap-x-2 rounded-lg p-1 break-all `, className)}>
       {children}
-      <button
-        className={twMerge(`text-green-400`, btnClassName)}
-        onClick={handleCopyToClipboard}
-      >
-        <BiSolidCopyAlt
-          size={21}
-          className="text-green-400 hover:text-green-300"
-        />
+      <button className={twMerge(`text-green-400`, btnClassName)} onClick={handleCopyToClipboard}>
+        <BiSolidCopyAlt size={21} className="text-green-400 hover:text-green-300" />
       </button>
     </div>
   );
